requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.2
Django==3.2.19
django-appconf==1.0.5
django-datetimepicker==3.14
django-user-accounts==3.2.0
gunicorn==21.2.0
importlib-metadata==6.7.0
packaging==23.1
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
whitenoise==6.5.0
zipp==3.15.0
